# perf: improve setDottedPath