--- a/lib/helpers/path/setDottedPath.js
+++ b/lib/helpers/path/setDottedPath.js
@@ -1,9 +1,14 @@
 'use strict';
 
 module.exports = function setDottedPath(obj, path, val) {
-  const parts = path.indexOf('.') === -1 ? [path] : path.split('.');
+  if (path.indexOf('.') === -1) {
+    obj[path] = val;
+    return;
+  }
+  const parts = path.split('.');
+  const last = parts.pop();
   let cur = obj;
-  for (const part of parts.slice(0, -1)) {
+  for (const part of parts) {
     if (cur[part] == null) {
       cur[part] = {};
     }
@@ -11,6 +16,5 @@ module.exports = function setDottedPath(obj, path, val) {
     cur = cur[part];
   }
 
-  const last = parts[parts.length - 1];
   cur[last] = val;
 };