@vkarpov15 2022-01-24
It cannot unfortunately, because mpath explicitly avoids creating intermediate paths. So mpath setting `a.b.c.d` won't create intermediate `a.b`, `a.c` if `a.b` is undefined.

In theory, we could add that to mpath somehow. But this function exists specifically to work around that limitation in mpath.